Fix undefined behavior when multiple lines match
When there are multiple lines matched for the same key we could end up
with different results depending on the order that the lines are
evaluated.

From now on it is only accepted multiple matches for the same key if
they are not ambiguous, i.e. one is a superset of the other.