Account for remote url notation difference
Since 'remote.origin.url' for the main repository is set on initial
cloning and NOT on jobs_done.yml, it's possible to have a remote url that
refers to 'ssh://example.com/repo.git' while git's remote.origin.url
refers to 'ssh://example.com/repo' (which is the same repo). This caused
the job creation to fail because we do basic string comparison.

EDEN-1873